# Change ARRAY_INTERSECT to use TypedSet