--- a/presto-main/src/main/java/com/facebook/presto/operator/scalar/ArrayIntersectFunction.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/scalar/ArrayIntersectFunction.java
@@ -13,34 +13,21 @@
  */
 package com.facebook.presto.operator.scalar;
 
+import com.facebook.presto.operator.aggregation.TypedSet;
 import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.function.Description;
-import com.facebook.presto.spi.function.OperatorDependency;
 import com.facebook.presto.spi.function.ScalarFunction;
 import com.facebook.presto.spi.function.SqlType;
 import com.facebook.presto.spi.function.TypeParameter;
-import com.facebook.presto.spi.type.StandardTypes;
 import com.facebook.presto.spi.type.Type;
-import com.facebook.presto.type.TypeUtils;
 import com.google.common.collect.ImmutableList;
-import it.unimi.dsi.fastutil.ints.AbstractIntComparator;
-import it.unimi.dsi.fastutil.ints.IntArrays;
-import it.unimi.dsi.fastutil.ints.IntComparator;
-
-import java.lang.invoke.MethodHandle;
-
-import static com.facebook.presto.spi.function.OperatorType.LESS_THAN;
 
 @ScalarFunction("array_intersect")
 @Description("Intersects elements of the two given arrays")
 public final class ArrayIntersectFunction
 {
-    private static final int INITIAL_SIZE = 128;
-
-    private int[] leftPositions = new int[INITIAL_SIZE];
-    private int[] rightPositions = new int[INITIAL_SIZE];
     private final PageBuilder pageBuilder;
 
     @TypeParameter("E")
@@ -49,103 +36,47 @@ public final class ArrayIntersectFunction
         pageBuilder = new PageBuilder(ImmutableList.of(elementType));
     }
 
-    private static int compareBlockValue(Type type, Block leftBlock, int left, Block rightBlock, int right)
-    {
-        if (leftBlock.isNull(left) && rightBlock.isNull(right)) {
-            return 0;
-        }
-        if (leftBlock.isNull(left)) {
-            return -1;
-        }
-        if (rightBlock.isNull(right)) {
-            return 1;
-        }
-        return type.compareTo(leftBlock, left, rightBlock, right);
-    }
-
-    private static IntComparator intBlockCompare(Type type, Block block)
-    {
-        return new AbstractIntComparator()
-        {
-            @Override
-            public int compare(int left, int right)
-            {
-                return compareBlockValue(type, block, left, block, right);
-            }
-        };
-    }
-
     @TypeParameter("E")
     @SqlType("array(E)")
     public Block intersect(
-            @OperatorDependency(operator = LESS_THAN, returnType = StandardTypes.BOOLEAN, argumentTypes = {"E", "E"}) MethodHandle lessThanFunction,
             @TypeParameter("E") Type type,
             @SqlType("array(E)") Block leftArray,
             @SqlType("array(E)") Block rightArray)
     {
+        if (leftArray.getPositionCount() < rightArray.getPositionCount()) {
+            Block tempArray = leftArray;
+            leftArray = rightArray;
+            rightArray = tempArray;
+        }
+
         int leftPositionCount = leftArray.getPositionCount();
         int rightPositionCount = rightArray.getPositionCount();
 
-        if (leftPositionCount == 0) {
-            return leftArray;
-        }
         if (rightPositionCount == 0) {
             return rightArray;
-        }
-
-        if (leftPositions.length < leftPositionCount) {
-            leftPositions = new int[leftPositionCount];
-        }
-
-        if (rightPositions.length < rightPositionCount) {
-            rightPositions = new int[rightPositionCount];
         }
 
         if (pageBuilder.isFull()) {
             pageBuilder.reset();
         }
 
-        for (int i = 0; i < leftPositionCount; i++) {
-            leftPositions[i] = i;
-        }
+        TypedSet rightTypedSet = new TypedSet(type, rightPositionCount, "array_intersect");
         for (int i = 0; i < rightPositionCount; i++) {
-            rightPositions[i] = i;
+            rightTypedSet.add(rightArray, i);
         }
-        IntArrays.quickSort(leftPositions, 0, leftPositionCount, intBlockCompare(type, leftArray));
-        IntArrays.quickSort(rightPositions, 0, rightPositionCount, intBlockCompare(type, rightArray));
 
-        BlockBuilder resultBlockBuilder = pageBuilder.getBlockBuilder(0);
+        BlockBuilder blockBuilder = pageBuilder.getBlockBuilder(0);
 
-        int leftCurrentPosition = 0;
-        int rightCurrentPosition = 0;
-        int leftBasePosition;
-        int rightBasePosition;
-        int totalCount = 0;
-
-        while (leftCurrentPosition < leftPositionCount && rightCurrentPosition < rightPositionCount) {
-            leftBasePosition = leftCurrentPosition;
-            rightBasePosition = rightCurrentPosition;
-            int compareValue = compareBlockValue(type, leftArray, leftPositions[leftCurrentPosition], rightArray, rightPositions[rightCurrentPosition]);
-            if (compareValue > 0) {
-                rightCurrentPosition++;
-            }
-            else if (compareValue < 0) {
-                leftCurrentPosition++;
-            }
-            else {
-                type.appendTo(leftArray, leftPositions[leftCurrentPosition], resultBlockBuilder);
-                leftCurrentPosition++;
-                rightCurrentPosition++;
-                totalCount++;
-                while (leftCurrentPosition < leftPositionCount && TypeUtils.positionEqualsPosition(type, leftArray, leftPositions[leftBasePosition], leftArray, leftPositions[leftCurrentPosition])) {
-                    leftCurrentPosition++;
-                }
-                while (rightCurrentPosition < rightPositionCount && TypeUtils.positionEqualsPosition(type, rightArray, rightPositions[rightBasePosition], rightArray, rightPositions[rightCurrentPosition])) {
-                    rightCurrentPosition++;
-                }
+        // The intersected set can have at most rightPositionCount elements
+        TypedSet intersectTypedSet = new TypedSet(type, blockBuilder, rightPositionCount, "array_intersect");
+        for (int i = 0; i < leftPositionCount; i++) {
+            if (rightTypedSet.contains(leftArray, i)) {
+                intersectTypedSet.add(leftArray, i);
             }
         }
-        pageBuilder.declarePositions(totalCount);
-        return resultBlockBuilder.getRegion(resultBlockBuilder.getPositionCount() - totalCount, totalCount);
+
+        pageBuilder.declarePositions(intersectTypedSet.size());
+
+        return blockBuilder.getRegion(blockBuilder.getPositionCount() - intersectTypedSet.size(), intersectTypedSet.size());
     }
 }