--- a/presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestTypedSet.java
+++ b/presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestTypedSet.java
@@ -13,6 +13,7 @@
  */
 package com.facebook.presto.operator.aggregation;
 
+import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
@@ -33,6 +34,7 @@ import static io.airlift.slice.Slices.utf8Slice;
 import static java.util.Collections.nCopies;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
 public class TestTypedSet
@@ -67,40 +69,134 @@ public class TestTypedSet
     public void testGetElementPosition()
     {
         int elementCount = 100;
-        TypedSet typedSet = new TypedSet(BIGINT, elementCount, FUNCTION_NAME);
+        // Set initialTypedSetEntryCount to a small number to trigger rehash()
+        int initialTypedSetEntryCount = 10;
+        TypedSet typedSet = new TypedSet(BIGINT, initialTypedSetEntryCount, FUNCTION_NAME);
         BlockBuilder blockBuilder = BIGINT.createFixedSizeBlockBuilder(elementCount);
         for (int i = 0; i < elementCount; i++) {
             BIGINT.writeLong(blockBuilder, i);
             typedSet.add(blockBuilder, i);
         }
+
+        assertEquals(typedSet.size(), elementCount);
+
         for (int j = 0; j < blockBuilder.getPositionCount(); j++) {
             assertEquals(typedSet.positionOf(blockBuilder, j), j);
         }
     }
 
     @Test
-    public void testGetElementPositionRandom()
+    public void testGetElementPositionWithNull()
     {
-        BlockBuilder keys = VARCHAR.createBlockBuilder(null, 5);
-        VARCHAR.writeSlice(keys, utf8Slice("hello"));
-        VARCHAR.writeSlice(keys, utf8Slice("bye"));
-        VARCHAR.writeSlice(keys, utf8Slice("abc"));
-
-        TypedSet set = new TypedSet(VARCHAR, keys.getPositionCount(), FUNCTION_NAME);
-        for (int i = 0; i < keys.getPositionCount(); i++) {
-            set.add(keys, i);
+        int elementCount = 100;
+        // Set initialTypedSetEntryCount to a small number to trigger rehash()
+        int initialTypedSetEntryCount = 10;
+        TypedSet typedSet = new TypedSet(BIGINT, initialTypedSetEntryCount, FUNCTION_NAME);
+        BlockBuilder blockBuilder = BIGINT.createFixedSizeBlockBuilder(elementCount);
+        for (int i = 0; i < elementCount; i++) {
+            if (i % 10 == 0) {
+                blockBuilder.appendNull();
+            }
+            else {
+                BIGINT.writeLong(blockBuilder, i);
+            }
+            typedSet.add(blockBuilder, i);
         }
 
-        BlockBuilder values = VARCHAR.createBlockBuilder(null, 5);
-        VARCHAR.writeSlice(values, utf8Slice("bye"));
-        VARCHAR.writeSlice(values, utf8Slice("abc"));
-        VARCHAR.writeSlice(values, utf8Slice("hello"));
-        VARCHAR.writeSlice(values, utf8Slice("bad"));
+        // The internal elementBlock and hashtable of the typedSet should contain
+        // all distinct non-null elements plus one null
+        assertEquals(typedSet.size(), elementCount - elementCount / 10 + 1);
 
-        assertEquals(set.positionOf(values, 2), 0);
-        assertEquals(set.positionOf(values, 1), 2);
-        assertEquals(set.positionOf(values, 0), 1);
-        assertFalse(set.contains(values, 3));
+        int nullCount = 0;
+        for (int j = 0; j < blockBuilder.getPositionCount(); j++) {
+            // The null is only added to typedSet once, so the internal elementBlock subscript is shifted by nullCountMinusOne
+            if (!blockBuilder.isNull(j)) {
+                assertEquals(typedSet.positionOf(blockBuilder, j), j - nullCount + 1);
+            }
+            else {
+                // The first null added to typedSet is at position 0
+                assertEquals(typedSet.positionOf(blockBuilder, j), 0);
+                nullCount++;
+            }
+        }
+    }
+
+    @Test
+    public void testGetElementPositionWithProvidedEmptyBlockBuilder()
+    {
+        int elementCount = 100;
+        // Set initialTypedSetEntryCount to a small number to trigger rehash()
+        int initialTypedSetEntryCount = 10;
+
+        BlockBuilder emptyBlockBuilder = BIGINT.createFixedSizeBlockBuilder(elementCount);
+        TypedSet typedSet = new TypedSet(BIGINT, emptyBlockBuilder, initialTypedSetEntryCount, FUNCTION_NAME);
+        BlockBuilder externalBlockBuilder = BIGINT.createFixedSizeBlockBuilder(elementCount);
+        for (int i = 0; i < elementCount; i++) {
+            if (i % 10 == 0) {
+                externalBlockBuilder.appendNull();
+            }
+            else {
+                BIGINT.writeLong(externalBlockBuilder, i);
+            }
+            typedSet.add(externalBlockBuilder, i);
+        }
+
+        assertEquals(typedSet.size(), emptyBlockBuilder.getPositionCount());
+        assertEquals(typedSet.size(), elementCount - elementCount / 10 + 1);
+
+        for (int j = 0; j < typedSet.size(); j++) {
+            assertEquals(typedSet.positionOf(emptyBlockBuilder, j), j);
+        }
+    }
+
+    @Test
+    public void testGetElementPositionWithProvidedNonEmptyBlockBuilder()
+    {
+        int elementCount = 100;
+        // Set initialTypedSetEntryCount to a small number to trigger rehash()
+        int initialTypedSetEntryCount = 10;
+
+        PageBuilder pageBuilder = new PageBuilder(ImmutableList.of(BIGINT));
+        BlockBuilder firstBlockBuilder = pageBuilder.getBlockBuilder(0);
+
+        for (int i = 0; i < elementCount; i++) {
+            BIGINT.writeLong(firstBlockBuilder, i);
+        }
+        pageBuilder.declarePositions(elementCount);
+
+        // The secondBlockBuilder should already have elementCount rows.
+        BlockBuilder secondBlockBuilder = pageBuilder.getBlockBuilder(0);
+
+        TypedSet typedSet = new TypedSet(BIGINT, secondBlockBuilder, initialTypedSetEntryCount, FUNCTION_NAME);
+        BlockBuilder externalBlockBuilder = BIGINT.createFixedSizeBlockBuilder(elementCount);
+        for (int i = 0; i < elementCount; i++) {
+            if (i % 10 == 0) {
+                externalBlockBuilder.appendNull();
+            }
+            else {
+                BIGINT.writeLong(externalBlockBuilder, i);
+            }
+            typedSet.add(externalBlockBuilder, i);
+        }
+
+        assertEquals(typedSet.size(), secondBlockBuilder.getPositionCount() - elementCount);
+        assertEquals(typedSet.size(), elementCount - elementCount / 10 + 1);
+
+        for (int i = 0; i < typedSet.size(); i++) {
+            int expectedPositionInSecondBlockBuilder = i + elementCount;
+            assertEquals(typedSet.positionOf(secondBlockBuilder, expectedPositionInSecondBlockBuilder), expectedPositionInSecondBlockBuilder);
+        }
+    }
+
+    @Test
+    public void testGetElementPositionRandom()
+    {
+        TypedSet set = new TypedSet(VARCHAR, 1, FUNCTION_NAME);
+        testGetElementPositionRandomFor(set);
+
+        BlockBuilder emptyBlockBuilder = VARCHAR.createBlockBuilder(null, 3);
+        TypedSet setWithPassedInBuilder = new TypedSet(VARCHAR, emptyBlockBuilder, 1, FUNCTION_NAME);
+        testGetElementPositionRandomFor(setWithPassedInBuilder);
     }
 
     @Test
@@ -145,9 +241,46 @@ public class TestTypedSet
         }
     }
 
+    private void testGetElementPositionRandomFor(TypedSet set)
+    {
+        BlockBuilder keys = VARCHAR.createBlockBuilder(null, 5);
+        VARCHAR.writeSlice(keys, utf8Slice("hello"));
+        VARCHAR.writeSlice(keys, utf8Slice("bye"));
+        VARCHAR.writeSlice(keys, utf8Slice("abc"));
+
+        for (int i = 0; i < keys.getPositionCount(); i++) {
+            set.add(keys, i);
+        }
+
+        BlockBuilder values = VARCHAR.createBlockBuilder(null, 5);
+        VARCHAR.writeSlice(values, utf8Slice("bye"));
+        VARCHAR.writeSlice(values, utf8Slice("abc"));
+        VARCHAR.writeSlice(values, utf8Slice("hello"));
+        VARCHAR.writeSlice(values, utf8Slice("bad"));
+        values.appendNull();
+
+        assertEquals(set.positionOf(values, 4), -1);
+        assertEquals(set.positionOf(values, 2), 0);
+        assertEquals(set.positionOf(values, 1), 2);
+        assertEquals(set.positionOf(values, 0), 1);
+        assertFalse(set.contains(values, 3));
+
+        set.add(values, 4);
+        assertTrue(set.contains(values, 4));
+    }
+
     private static void testBigint(Block longBlock, int expectedSetSize)
     {
         TypedSet typedSet = new TypedSet(BIGINT, expectedSetSize, FUNCTION_NAME);
+        testBigintFor(typedSet, longBlock);
+
+        BlockBuilder emptyBlockBuilder = BIGINT.createBlockBuilder(null, expectedSetSize);
+        TypedSet typedSetWithPassedInBuilder = new TypedSet(BIGINT, emptyBlockBuilder, expectedSetSize, FUNCTION_NAME);
+        testBigintFor(typedSetWithPassedInBuilder, longBlock);
+    }
+
+    private static void testBigintFor(TypedSet typedSet, Block longBlock)
+    {
         Set<Long> set = new HashSet<>();
         for (int blockPosition = 0; blockPosition < longBlock.getPositionCount(); blockPosition++) {
             long number = BIGINT.getLong(longBlock, blockPosition);