Resolves:
https://github.com/prestodb/presto/issues/11493

This commit improves the performance comparing to the original sort-merge
solution. The gain is from 2.3x to 11x for different data types.

before

```
Benchmark                                        (name)   (type)  Mode  Cnt    Score   Error  Units
BenchmarkArrayIntersect.arrayIntersect  array_intersect   BIGINT  avgt   20  193.023 ± 2.840  ns/op
BenchmarkArrayIntersect.arrayIntersect  array_intersect  VARCHAR  avgt   20  336.516 ± 5.301  ns/op
BenchmarkArrayIntersect.arrayIntersect  array_intersect   DOUBLE  avgt   20  219.812 ± 6.615  ns/op
BenchmarkArrayIntersect.arrayIntersect  array_intersect  BOOLEAN  avgt   20   23.913 ± 0.553  ns/op
```

After

```
Benchmark                                (type)  Mode  Cnt    Score   Error  Units
BenchmarkArrayIntersect.arrayIntersect   BIGINT  avgt   20   30.175 ± 5.077  ns/op
BenchmarkArrayIntersect.arrayIntersect  VARCHAR  avgt   20  146.365 ± 5.405  ns/op
BenchmarkArrayIntersect.arrayIntersect   DOUBLE  avgt   20   63.752 ± 5.329  ns/op
BenchmarkArrayIntersect.arrayIntersect  BOOLEAN  avgt   20    2.133 ± 0.243  ns/op
```