@yingsu00 2018-12-06
I did some JMH benchmarks on the two possible changes to TypedSet:

	1) Instead of building an elementBlock inside of TypdeSet, we can directly pass in the block and only build a hashtable of block positions. This way we avoid one data copy and Block building, and also removed the 4MB limit on the elementBlock Size. However as @haozhun pointed out this may have memory GC issue for aggregatoin functions like MultimapAggregationFunction , because this way we have to hold the input block longer until we're done with the TypedSet, while the other array/map functions don't matter that much because the life cycle of the input blocks and TypedSet are not muh different. But since there're only one aggregation function, this can probably work.
	Another change we need to make this work is to be able to add more than one block to the TypedSet. There are a few use cases(listed in the table below) requires this, e.g. ArrayExceptFunction. This is doable, we just need to keep track of the added blocks inside of TypedSet, and maybe use some additional internal hashtable to find which block the value is in. There is a space penalty for this of course, but because we don't need to build the internal elementBlock the space requirement should not be more than existing implementation.
	
	2) There're a number of use cases the internal elementBlock can be DIRECTLY returned and used as results. Such cases are also listed in the table below. Theoretically this would save some cost of building the a separate result blockBuilder in caller(e.g. distinctElementBlockBuilder in ArrayDistinctFunction), however in the JMH benchmarks I'm not seeing any obvious performance gains, except that this should save memory usage any ways. If we do 1) we can also make TypdeSet to return a result block based on an array of positions, but again this doesn't show any performance gain/loss comparing to using a separate result block builder in the caller.


This table lists all usages of TypdedSet. We can see 5 of them requires the ability to add more than one Blocks so this feature is needed. Also there're 6 callers that can directly use the current TypedSet internal elementBlock as result. However if we use approach 1) and just pass in the input block, the benefit of saving memory usage by just returning the elementBlock directly goes away.

```
  | Aggregation Function | Input Blocks | Handle > 1 blocks add | Can directly use the elementBlock as output
-- | -- | -- | -- | --
MultimapAggregationFunction | Y | 1 | N | Y
ArrayDistinctFunction | N | 1 | N | Y
ArrayExceptFunction | N | 2 | Y | N
ArrayIntersectFunction (Using hashtable solution) | N | 2 | N | Y
ArrayUnionFunction | N | 2 | Y | Y
MapConcatFunction | N | n | Y | Y
MapFromEntriesFunction | N | 1   N SingleRowBlock | Y | Y (for keyBlock)
MapToMapCast | N | 1 | N | Y(need to apply valueProcessFunction.invokeExact)
mapDotProduct | N | 2 | N | N
MultimapFromEntriesFunction | N | 1 | N | N
```
Without knowing the existence of JsonUtil#HashTable, I implemented a class NoCopyTypedSet and did a few benchmarks on array_distinct. The logic is similar to JsonUtil#HashTable, except NoCopyTypedSet handles nulls while JsonUtil#HashTable doesnt. If we use JsonUtil#HashTable directly to implement functions like array_distinct it would give wrong results on 0 and null.  As a result NoCopyTypedSet is slightly slower than HashTable for about 10%. The tests were done on arrays with 10000 varchar values.

The first test shows directly returning elementBlock doesn't have much gain in performance:

TypedSet, caller build the result Block in a separrate block builder:
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20  35.440 ± 2.055  ns/op

TypedSet, directly return elementBlock
BenchmarkArrayDistinct.arrayDistinct  array_distinct  avgt   20  34.890 ± 0.916  ns/op


The next test using NoCopyTypedSet compares 1) return a block in NoCopyTypedSet .getBlock() vs 2)the caller uses a separate block builder:

NoCopyTypedSet  return the result Block in NoCopyTypedSet.getBlock() by calling Block.getPositions():
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20  29.871 ± 2.498  ns/op

NoCopyTYpedSet, caller use separate blockBuilder
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20  28.951 ± 1.027  ns/op


Again there's no obvious difference. NoCopyTypedSet  is faster than TypedSet because I changed the internal hashtable to int[] instead of IntArrayList.

The JsonUtil#HashTable results;
JsonUtil#HashTable,  return the result Block in NoCopyTypedSet.getBlock()
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20  26.472 ± 0.411  ns/op

JsonUtil#HashTable, caller use separate blockBuilder
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20  27.062 ± 0.794  ns/op

Based on the above, I think the best way is to use approach 1), ie. Make JsonUtil#HashTable  a top level class but make it a) be able to handle nulls correctly b) be able to add > 1 blocks. This way we can remove the 4MB limit which users complaint about. MultimapAggregationFunction can still use original TypedSet if we observe the memory pressure. I can make this work for array_intersect first and watch for any performance penalty for adding > 1 blocks.  @dain @wenleix @haozhun what do you think?