@wenleix 2018-12-11
The benchmark result at top looks great. Note the bigger input arrays are, the better improvements we will see. Thus can we look at the production workload, and generate benchmark with input array of that size ? Given the stack depth of the frame-chart , I suspect the speed-up is likely closer to the upper side in your benchmark.