@wenleix 2018-12-11
Given the complicated trade-offs between different proposals on `TypedSet`, I would suggest taking one step back and focusing on impact  -- switch from sort based solution to a hash based solution.  

The blocking issue (https://github.com/prestodb/presto/pull/11984#pullrequestreview-178966684) is to avoid new `BlockBuilder` created by `TypedSet` per each invocation. We can just change `TypedSet` to allow a provided container `BlockBuilder` to achieve this goal. All the follow-up optimizations and proposals are great, but let's focus on impact and get rid of sort-based `array_intersect` first ;) . Does that make sense? @haozhun , @yingsu00 

Also note, for heavy use cases, we can always use specialized implementation (For example, we can use `LongSet` instead of `TypedSet` for `BIGINT`)